[#35] Block要素から文字列を得る処理を共通モジュールへ切り出し
Block要素から文字列を得る処理は、converterだけでなく、builderからも参照される
※hタグを組み立てるときにID属性のテキストを設定するのに参照
責務の異なるモジュール間で参照されるモジュールをどちらか一方に置いておくと、境界が曖昧になってしまう
各モジュールを独立させられるよう、共通処理は別で切り出した
