+++ a_pompom_markdown_parser/converter/toc_converter.py
@@ -1,26 +1,9 @@
 import dataclasses
 from typing import Generator
 
-from a_pompom_markdown_parser.element.block import ParseResult, Block, HeadingBlock, ListBlock, ListItemBlock
+from a_pompom_markdown_parser.element.block import ParseResult, HeadingBlock, ListBlock, ListItemBlock
 from a_pompom_markdown_parser.element.inline import LinkInline
-
-
-def get_text_from_block(block: Block) -> str:
-    """
-    Inline要素を子に含むBlock要素から、マークダウンの記法を除いたテキストを抽出
-
-    :param block: 対象Block要素
-    :return: Block要素が含むマークダウンの記法を除いた文字列
-    """
-
-    text = ''
-    for child in block.children:
-        if isinstance(child, Block):
-            text += get_text_from_block(child)
-
-        text += child.text
-
-    return text
+from a_pompom_markdown_parser.block_utility import get_text_from_block
 
 
 @dataclasses.dataclass
